refactor(frontend): move componentes para atoms, molecules, organisms e templates
