+++ frontend/src/components/atoms/icons/CommunityIcon.vue
