+++ frontend/src/components/atoms/icons/EventIcon.vue
