+++ frontend/src/components/atoms/icons/HomeIcon.vue
