+++ frontend/src/components/atoms/icons/LogoutIcon.vue
