+++ frontend/src/components/atoms/icons/MapIcon.vue
