+++ frontend/src/components/atoms/icons/RiverIcon.vue
