+++ frontend/src/components/molecules/SidebarProfileCard.vue
