+++ frontend/src/components/organisms/LeftAuthSide.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import Card from './Card.vue';
+import Card from '@/components/molecules/Card.vue';
 
 </script>
