+++ frontend/src/views/HomeView.vue
@@ -23,7 +23,7 @@
 </template>
 
 <script setup lang="ts">
-import BaseButton from '@/components/base/BaseButton.vue';
+import BaseButton from '@/components/atoms/BaseButton.vue';
 
 
 </script>
